+++ src/main/java/com/gabriel/SpringEcom/dto/CartDTO/CartItemRequestDTO.java
@@ -1,6 +1,6 @@
 package com.gabriel.SpringEcom.dto.CartDTO;
 
 public record CartItemRequestDTO(
-        Integer productId,
+        Long productId,
         int quantity
 ) {}
