+++ src/main/java/com/gabriel/SpringEcom/dto/OrderDTO/OrderItemRequest.java
@@ -1,6 +1,6 @@
 package com.gabriel.SpringEcom.dto.OrderDTO;
 
 public record OrderItemRequest(
-        int productId,
+        Long productId,
         int quantity
 ) {}
